packages/cloud_api/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
orjson==3.10.7
sqlmodel==0.0.22
sqlite-utils==3.37


